fix: Permission for API > 30 and picture for f2hwg prediction

+++ app/src/main/java/com/kalorize/kalorizeappmobile/data/remote/response/F2hwgResponse.kt
@@ -7,8 +7,8 @@ data class F2hwgResponse(
     val gender: String,
 
     @SerializedName("height")
-    val height: Long,
+    val height: Float,
 
     @SerializedName("weight")
-    val weight: Long
+    val weight: Float
 )
